# crates/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DisconnectedResponse": {
        "properties": {
          "disconnected": {
            "type": "boolean"
          }
        },
        "required": [
          "disconnected"
        ],
        "type": "object"
      },
...
      "JiraBoard": {
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type"
        ],
        "type": "object"
      },
...
      "JiraCreateMeta": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/JiraFieldDescriptor"
            },
            "type": "array"
          },
          "issueTypeId": {
            "type": "string"
          }
        },
        "required": [
          "issueTypeId",
          "fields"
        ],
        "type": "object"
      },
...
      "JiraEditMeta": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/JiraFieldDescriptor"
            },
            "type": "array"
          }
        },
        "required": [
          "fields"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Workflow",
      "url": "https://github.com/"
    },
...
    "license": {
      "name": "Proprietary"
    },
    "title": "Workflow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me/jira": {
      "delete": {
        "operationId": "jiraDisconnect",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisconnectedResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "DELETE /me/jira — disconnect.",
        "tags": [
          "jira"
        ]
      },
      "get": {
        "operationId": "jiraStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraConnectionSummary"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/jira — connection summary.",
        "tags": [
          "jira"
        ]
      }
    },
...
    "/api/me/jira/boards": {
      "get": {
        "operationId": "jiraBoards",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JiraBoard"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/jira/boards — agile boards.",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/createmeta": {
      "get": {
        "operationId": "jiraCreateMeta",
        "parameters": [
          {
            "in": "query",
            "name": "projectKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "issueTypeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraCreateMeta"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/jira/createmeta?projectKey&issueTypeId — create field metadata.",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/dashboard": {
      "get": {
        "operationId": "jiraDashboard",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraDashboard"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/jira/dashboard — multi-queue dashboard.",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/editmeta": {
      "get": {
        "operationId": "jiraEditMeta",
        "parameters": [
          {
            "description": "Issue key",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraEditMeta"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/jira/editmeta?key — edit field metadata.",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/issue": {
      "get": {
        "operationId": "jiraIssue",
        "parameters": [
          {
            "description": "Issue key",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraIssueDetail"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/jira/issue?key — issue detail.",
        "tags": [
          "jira"
        ]
      },
      "post": {
        "operationId": "jiraCreateIssue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JiraCreateIssueInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraCreateIssueResult"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "POST /me/jira/issue — create an issue (fields pass the metadata allowlist).",
        "tags": [
          "jira"
        ]
      },
      "put": {
        "operationId": "jiraEditIssue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "PUT /me/jira/issue — edit an issue.",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/issue/assign": {
      "post": {
        "operationId": "jiraAssign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "POST /me/jira/issue/assign — (re)assign or unassign (null).",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/issue/comment": {
      "post": {
        "operationId": "jiraComment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraComment"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "POST /me/jira/issue/comment — add a comment.",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/issue/transition": {
      "post": {
        "operationId": "jiraTransition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "POST /me/jira/issue/transition — apply a transition.",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/issue/transitions": {
      "get": {
        "operationId": "jiraTransitions",
        "parameters": [
          {
            "description": "Issue key",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JiraTransition"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/jira/issue/transitions?key — available transitions.",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/issue/worklog": {
      "post": {
        "operationId": "jiraWorklog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorklogBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "POST /me/jira/issue/worklog — log work.",
        "tags": [
          "jira"
        ]
      }
    },
    "/api/me/jira/issuetypes": {
      "get": {
        "operationId": "jiraIssueTypes",
        "parameters": [
          {
            "in": "query",
            "name": "projectKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JiraIssueType"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/jira/issuetypes?projectKey — issue types for a project.",
        "tags": [
          "jira"
        ]
      }
    },
...
    "/api/me/jira/queue": {
      "get": {
        "operationId": "jiraQueue",
        "parameters": [
          {
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraIssuePage"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/jira/queue?key&cursor — one queue's issue page.",
        "tags": [
          "jira"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Same-origin (the web app proxies /api to this server)",
      "url": "/"
    }
  ],
...
}